cs_backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.2,<26.0
django-redis>=5.4,<6.0
mysqlclient>=2.2,<3.0
python-dotenv>=1.0,<2.0
